compass-api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.5.3
brotlipy==0.6.0
cffi==1.9.1               # via brotlipy
click==6.7                # via rq
contextlib2==0.5.4        # via raven
coreapi==2.3.0            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
diff-match-patch==20121119  # via django-import-export
django-cors-middleware==1.3.1
django-environ==0.4.1
django-filter==1.0.1
django-haystack==2.6.0    # via drf-haystack
django-import-export==0.5.1
django-rest-swagger==2.1.2
django-rq==0.9.5
Django==1.10.6            # via django-environ, django-haystack, django-rq, drf-haystack
djangorestframework==3.6.2
drf-haystack==1.6.1
elasticsearch==1.9.0
grako==3.19.1
gunicorn==19.7.0
itypes==1.1.0             # via coreapi
jinja2==2.9.5             # via coreschema
MarkupSafe==1.0           # via jinja2
netaddr==0.7.19
openapi-codec==1.3.1      # via django-rest-swagger
psycopg2==2.7.1
pycparser==2.17           # via cffi
python-dateutil==2.6.0    # via drf-haystack
raven==6.0.0
redis==2.10.5             # via rq
requests==2.13.0          # via coreapi
rq==0.7.1                 # via django-rq
schedule==0.4.2
simplejson==3.10.0        # via django-rest-swagger
six==1.10.0               # via django-environ, python-dateutil
tablib==0.11.4            # via django-import-export
uritemplate==3.0.0        # via coreapi
urllib3==1.20             # via elasticsearch
uwsgi==2.0.14
